dealing with roundings

+++ src/main/java/ru/xxd3c0yxx/nf_test_1/ICalculator.java
@@ -1,8 +1,9 @@
 package ru.xxd3c0yxx.nf_test_1;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public interface ICalculator {
-    VacPayment calculateVacationPay(double wage, long vacationDays);
+    VacPayment calculateVacationPay(BigDecimal wage, long vacationDays);
     long vacationLenCalc(Date vacationStart, int vacationDays);
 }
